[impl] move Config into its own file
Also implement initial structures for it to support #[repr(uN)] attribute.
