+++ impl/src/bitfield_attr.rs
@@ -1,114 +1,6 @@
-use crate::errors::CombineError;
 use core::convert::TryFrom;
-use proc_macro2::Span;
 use syn::spanned::Spanned;
-
-#[derive(Default)]
-pub struct Config {
-    pub specifier: Option<ConfigValue<bool>>,
-    pub bytes: Option<ConfigValue<usize>>,
-    pub filled: Option<ConfigValue<bool>>,
-}
-
-pub struct ConfigValue<T> {
-    pub value: T,
-    pub span: Span,
-}
-
-impl<T> ConfigValue<T> {
-    pub fn new(value: T, span: Span) -> Self {
-        Self { value, span }
-    }
-}
-
-impl Config {
-    /// Returns the value of the `specifier` parameter if provided and otherwise `false`.
-    pub fn specifier_enabled(&self) -> bool {
-        self.specifier
-            .as_ref()
-            .map(|config| config.value)
-            .unwrap_or(false)
-    }
-
-    /// Returns the value of the `filled` parameter if provided and otherwise `true`.
-    pub fn filled_enabled(&self) -> bool {
-        self.filled
-            .as_ref()
-            .map(|config| config.value)
-            .unwrap_or(true)
-    }
-}
-
-impl Config {
-    /// Sets the `specifier: bool` #[bitfield] parameter to the given value.
-    ///
-    /// # Errors
-    ///
-    /// If the specifier has already been set.
-    pub fn specifier(&mut self, value: bool, span: Span) -> Result<(), syn::Error> {
-        match &self.specifier {
-            Some(previous) => {
-                return Err(format_err!(
-                    span,
-                    "encountered duplicate `specifier` parameter: duplicate set to {:?}",
-                    previous.value
-                )
-                .into_combine(format_err!(
-                    previous.span,
-                    "previous `specifier` parameter here"
-                )))
-            }
-            None => self.specifier = Some(ConfigValue::new(value, span)),
-        }
-        Ok(())
-    }
-
-    /// Sets the `bytes: int` #[bitfield] parameter to the given value.
-    ///
-    /// # Errors
-    ///
-    /// If the specifier has already been set.
-    pub fn bytes(&mut self, value: usize, span: Span) -> Result<(), syn::Error> {
-        match &self.bytes {
-            Some(previous) => {
-                return Err(format_err!(
-                    span,
-                    "encountered duplicate `bytes` parameter: duplicate set to {:?}",
-                    previous.value
-                )
-                .into_combine(format_err!(
-                    previous.span,
-                    "previous `bytes` parameter here"
-                )))
-            }
-            None => self.bytes = Some(ConfigValue::new(value, span)),
-        }
-        Ok(())
-    }
-
-    /// Sets the `filled: bool` #[bitfield] parameter to the given value.
-    ///
-    /// # Errors
-    ///
-    /// If the specifier has already been set.
-    pub fn filled(&mut self, value: bool, span: Span) -> Result<(), syn::Error> {
-        match &self.filled {
-            Some(previous) => {
-                return Err(format_err!(
-                    span,
-                    "encountered duplicate `filled` parameter: duplicate set to {:?}",
-                    previous.value
-                )
-                .into_combine(format_err!(
-                    previous.span,
-                    "previous `filled` parameter here"
-                )))
-            }
-            None => self.filled = Some(ConfigValue::new(value, span)),
-        }
-        Ok(())
-    }
-}
+use crate::config::Config;
 
 /// Raises an unsupported argument compile time error.
 fn unsupported_argument<T>(arg: T) -> syn::Error
@@ -116,6 +8,7 @@ where
     T: Spanned,
 {
     format_err!(arg, "encountered unsupported #[bitfield] attribute")
+
 }
 
 /// The parameters given to the `#[bitfield]` proc. macro.
